Delete alert file. Create new files for buttons, constructor, createtodo, and localstorage. Begin process of breaking things apart.
