--- lib/alert.js
@@ -1,2 +1,0 @@
-
-module.exports = () => { alert('Test') }
