+++ lib/index.js
@@ -1,4 +1,11 @@
+const buttons = require('./buttons');
+const storeIdeas = require('./localstorage');
+const path = require('path');
+const ideaTemplate = require('./constructor');
+require('./script.js');
 
+buttons.saveButton();
+// storage.storeIdeas();
 
 
 
@@ -9,4 +16,3 @@
 
 
 // require('./styles.scss');
-require('./script.js');
