+++ lib/script.js
@@ -1,61 +1,20 @@
-var ideas = JSON.parse(localStorage.getItem('savedArray')) || [];
+const buttons = require('./buttons');
+const ideaTemplate = require('./constructor');
 
-$(document).ready(function() {
-  renderFromStorage();
-})
 
-function ideaTemplate(title, body, id, importance) {
-  this.title = title;
-  this.body = body;
-  this.id = Date.now();
-  this.importance = importance || 'normal';
-}
+// $(document).ready(function() {
+//   renderFromStorage();
+// })
 
-$('.save-button').on('click', function(e) {
-  e.preventDefault();
-  $('.todo-render').remove();
-  var $titleInput = $('.todo-title').val();
-  var $ideaInput = $('.todo-task').val();
-  var newIdea = new ideaTemplate($titleInput, $ideaInput);
-  ideas.push(newIdea);
-  storeIdeas();
-  renderFromStorage();
-  clearInputs();
-});
+// function renderFromStorage() {
+//   for (var i = 0; i < ideas.length; i++) {
+//     renderIdea(ideas[i]);
+//   }
+// };
 
-function storeIdeas() {
-  localStorage.setItem('savedArray', JSON.stringify(ideas));
-};
 
-function renderFromStorage() {
-  for (var i = 0; i < ideas.length; i++) {
-    renderIdea(ideas[i]);
-  }
-};
 
-function renderIdea(newIdea) {
-  $('.list-section').prepend(`
-    <div id=${newIdea.id} class ="todo-render">
-      <div class="id-wrapper">
-        <h2 class ="title-render" contenteditable="true">${newIdea.title}</h2>
-        <button class="button delete">delete</button>
-      </div>
-      <p class="editable-body" contenteditable="true">${newIdea.body}</p>
-      <div id="button-wrapper">
-        <div id="vote-buttons">
-        <button class="button upvote"></button>
-        <button class="button downvote"></button>
-        <label class="quality-text">importance: ${newIdea.importance}</label>
-      </div>
-        <button class="button done"></button>
-      </div>
-    </div>`)
-}
 
-function clearInputs() {
-  $('.todo-title').val('');
-  $('.todo-task').val('');
-}
 
 $('.list-section').on('click', '.done', function() {
   $(this).closest('.todo-render').toggleClass('task-completed');
